requirements.txt
# Core dependencies
flask>=2.3.0
pyyaml>=6.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Optional: Real adapter integrations (uncomment as needed)
# google-cloud-vision>=3.0.0        # For Google Vision OCR
# boto3>=1.26.0                     # For AWS services (Textract, SES, DynamoDB)
# pytesseract>=0.3.10               # For Tesseract OCR
# psycopg2-binary>=2.9.0            # For PostgreSQL
# requests>=2.28.0                  # For HTTP-based APIs (Clearbit, etc.)
# sendgrid>=6.9.0                   # For SendGrid email

